refactor: oiiotool break out expression eval methods into separate module
This is just a code move. There no changes to the code logic, though
in the process, I do change old "local" uses of eval_pool to newthe
Strutil one.
